+++ sdk/servicebus/azure-messaging-servicebus/src/main/java/com/azure/messaging/servicebus/administration/models/CorrelationRuleFilter.java
@@ -32,7 +32,7 @@ import java.util.Objects;
  * match.
  * <p>
  * This provides an efficient shortcut for declarations of filters that deal only with correlation
- * equality. In this case the cost of the lexigraphical analysis of the expression can be avoided. Not only will
+ * equality. In this case the cost of the lexicographical analysis of the expression can be avoided. Not only will
  * correlation filters be optimized at declaration time, but they will also be optimized at runtime. Correlation filter
  * matching can be reduced to a hashtable lookup, which aggregates the complexity of the set of defined correlation
  * filters to <code>O(1)</code>.