+++ sdk/servicebus/azure-messaging-servicebus/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Features Added
 - Added rule manager client to manage rules for ServiceBus subscription with listen claims. ([#27711](https://github.com/Azure/azure-sdk-for-java/issues/27711))
 - Added ability to create a subscription with default rule. ([#29885](https://github.com/Azure/azure-sdk-for-java/issues/29885))
-
+- `ServiceBusAdministrationClientBuilder` now supports using `AzureSasCredential`. ([#30255](https://github.com/Azure/azure-sdk-for-java/issues/30255))
 ### Breaking Changes
 
 ### Bugs Fixed
@@ -329,7 +329,7 @@ Fixed the issue that the second call of `ServiceBusReceiverClient.complete` is s
 
 ### Breaking Changes
 - Changed `receiveMessages` API to return `ServiceBusReceivedMessage` instead of ServiceBusReceivedMessageContext in 
-  `ServiceBusReceiverAsynClient` and `ServiceBusReceiverClient`.
+  `ServiceBusReceiverAsyncClient` and `ServiceBusReceiverClient`.
 - Removed `SendVia` option from `ServiceBusClientBuilder`. See issue for more detail 
   [16942](https://github.com/Azure/azure-sdk-for-java/pull/16942).
 - Removed `sessionId` setting from `ServiceBusSessionReceiverClientBuilder` as creating receiver clients bound to a 
@@ -339,7 +339,7 @@ Fixed the issue that the second call of `ServiceBusReceiverClient.complete` is s
   `ServiceBusSessionProcessorClientBuilder` as the feature of receiving messages from multiple sessions is moved from 
   the receiver client to the new `ServiceBusSessionProcessorClient`.
 - Renamed `tryAdd` to `tryAddMessage` in `ServiceBusMessageBatch`.
-- Removed `sessionId` specific methods from `ServiceBusReceiverAsynClient` and `ServiceBusReceiverClient` because now 
+- Removed `sessionId` specific methods from `ServiceBusReceiverAsyncClient` and `ServiceBusReceiverClient` because now 
   receiver client is always tied to one session. 
   
 ### Bug Fixes